refactor: Route Enum 수정 - ROUTE -> Route 로 클래스 이름 수정 - Route 내부의 값을 STATIC, TEMPLATE으로 두어 상대 경로로 동작하도록 설정
